feat(transport): move strings
