+++ transport/grpc/security/jwt/jwt.go
@@ -8,8 +8,8 @@ import (
 	"github.com/alexfalkowski/go-service/security/header"
 	sjwt "github.com/alexfalkowski/go-service/security/jwt"
 	"github.com/alexfalkowski/go-service/security/jwt/meta"
-	"github.com/alexfalkowski/go-service/strings"
 	smeta "github.com/alexfalkowski/go-service/transport/grpc/meta"
+	"github.com/alexfalkowski/go-service/transport/strings"
 	middleware "github.com/grpc-ecosystem/go-grpc-middleware"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
