+++ transport/http/trace/opentracing/opentracing.go
@@ -7,8 +7,8 @@ import (
 	"time"
 
 	"github.com/alexfalkowski/go-service/meta"
-	sstrings "github.com/alexfalkowski/go-service/strings"
 	"github.com/alexfalkowski/go-service/trace/opentracing"
+	tstrings "github.com/alexfalkowski/go-service/transport/strings"
 	"github.com/alexfalkowski/go-service/version"
 	otr "github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/ext"
@@ -55,7 +55,7 @@ func NewHandler(tracer Tracer, handler http.Handler) *Handler {
 // ServeHTTP for otr.
 func (h *Handler) ServeHTTP(resp http.ResponseWriter, req *http.Request) {
 	service, method := req.URL.Path, strings.ToLower(req.Method)
-	if sstrings.IsHealth(service) {
+	if tstrings.IsHealth(service) {
 		h.Handler.ServeHTTP(resp, req)
 
 		return
@@ -101,7 +101,7 @@ type RoundTripper struct {
 }
 
 func (r *RoundTripper) RoundTrip(req *http.Request) (*http.Response, error) {
-	if sstrings.IsHealth(req.URL.String()) {
+	if tstrings.IsHealth(req.URL.String()) {
 		return r.RoundTripper.RoundTrip(req)
 	}
